HHH-18620 - Add @NativeGenerator
