+++ hibernate-core/src/main/java/org/hibernate/boot/model/internal/GeneratorBinder.java
@@ -725,6 +725,7 @@ public class GeneratorBinder {
 					persistentClass,
 					idValue,
 					idMember,
+					generatedValue,
 					context
 			) );
 		}
