+++ hibernate-core/src/main/java/org/hibernate/boot/model/internal/StrictIdGeneratorResolverSecondPass.java
@@ -4,13 +4,8 @@
  */
 package org.hibernate.boot.model.internal;
 
-import java.util.HashMap;
-import java.util.Map;
 import java.util.UUID;
 
-import org.hibernate.MappingException;
-import org.hibernate.boot.model.IdentifierGeneratorDefinition;
-import org.hibernate.boot.model.relational.Database;
 import org.hibernate.boot.models.annotations.internal.SequenceGeneratorJpaAnnotation;
 import org.hibernate.boot.models.annotations.internal.TableGeneratorJpaAnnotation;
 import org.hibernate.boot.models.spi.GenericGeneratorRegistration;
@@ -19,31 +14,16 @@ import org.hibernate.boot.models.spi.SequenceGeneratorRegistration;
 import org.hibernate.boot.models.spi.TableGeneratorRegistration;
 import org.hibernate.boot.spi.InFlightMetadataCollector;
 import org.hibernate.boot.spi.MetadataBuildingContext;
-import org.hibernate.dialect.Dialect;
-import org.hibernate.generator.Generator;
-import org.hibernate.id.enhanced.SequenceStyleGenerator;
-import org.hibernate.mapping.Column;
 import org.hibernate.mapping.PersistentClass;
 import org.hibernate.mapping.SimpleValue;
 import org.hibernate.models.spi.MemberDetails;
-import org.hibernate.resource.beans.container.spi.BeanContainer;
 
 import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.SequenceGenerator;
 
-import static org.hibernate.boot.model.internal.GeneratorAnnotationHelper.*;
+import static org.hibernate.boot.model.internal.GeneratorAnnotationHelper.handleGenericGenerator;
+import static org.hibernate.boot.model.internal.GeneratorAnnotationHelper.handleSequenceGenerator;
+import static org.hibernate.boot.model.internal.GeneratorAnnotationHelper.handleTableGenerator;
 import static org.hibernate.boot.model.internal.GeneratorAnnotationHelper.handleUuidStrategy;
-import static org.hibernate.boot.model.internal.GeneratorParameters.fallbackAllocationSize;
-import static org.hibernate.boot.model.internal.GeneratorParameters.identityTablesString;
-import static org.hibernate.boot.model.internal.GeneratorStrategies.mapLegacyNamedGenerator;
-import static org.hibernate.id.IdentifierGenerator.ENTITY_NAME;
-import static org.hibernate.id.IdentifierGenerator.GENERATOR_NAME;
-import static org.hibernate.id.IdentifierGenerator.JPA_ENTITY_NAME;
-import static org.hibernate.id.OptimizableGenerator.IMPLICIT_NAME_BASE;
-import static org.hibernate.id.OptimizableGenerator.INCREMENT_PARAM;
-import static org.hibernate.id.PersistentIdentifierGenerator.PK;
-import static org.hibernate.id.PersistentIdentifierGenerator.TABLE;
-import static org.hibernate.id.PersistentIdentifierGenerator.TABLES;
 
 /**
  * SecondPass implementing delayed resolution of id-generators associated with an entity
@@ -60,46 +40,18 @@ import static org.hibernate.id.PersistentIdentifierGenerator.TABLES;
  *
  * @author Steve Ebersole
  */
-public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver {
-	private final PersistentClass entityMapping;
-	private final SimpleValue idValue;
-	private final MemberDetails idMember;
-
-	private final MetadataBuildingContext buildingContext;
-
+public class StrictIdGeneratorResolverSecondPass extends AbstractEntityIdGeneratorResolver {
 	public StrictIdGeneratorResolverSecondPass(
 			PersistentClass entityMapping,
 			SimpleValue idValue,
 			MemberDetails idMember,
+			GeneratedValue generatedValue,
 			MetadataBuildingContext buildingContext) {
-		this.entityMapping = entityMapping;
-		this.idValue = idValue;
-		this.idMember = idMember;
-		this.buildingContext = buildingContext;
+		super( entityMapping, idValue, idMember, generatedValue, buildingContext );
 	}
 
 	@Override
-	public void doSecondPass(Map<String, PersistentClass> persistentClasses) throws MappingException {
-		final GeneratedValue generatedValue = idMember.getDirectAnnotationUsage( GeneratedValue.class );
-		switch ( generatedValue.strategy() ) {
-			case UUID -> handleUuidStrategy( idValue, idMember, buildingContext );
-			case IDENTITY -> handleIdentityStrategy( idValue );
-			case SEQUENCE -> handleSequenceStrategy( generatedValue );
-			case TABLE -> handleTableStrategy( generatedValue );
-			case AUTO -> handleAutoStrategy( generatedValue );
-		}
-	}
-
-	private void handleSequenceStrategy(GeneratedValue generatedValue) {
-		if ( generatedValue.generator().isEmpty() ) {
-			handleUnnamedSequenceGenerator();
-		}
-		else {
-			handleNamedSequenceGenerator( generatedValue );
-		}
-	}
-
-	private void handleUnnamedSequenceGenerator() {
+	protected void handleUnnamedSequenceGenerator() {
 		final InFlightMetadataCollector metadataCollector = buildingContext.getMetadataCollector();
 
 		// according to the spec, this should locate a generator with the same name as the entity-name
@@ -126,7 +78,8 @@ public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver 
 		);
 	}
 
-	private void handleNamedSequenceGenerator(GeneratedValue generatedValue) {
+	@Override
+	protected void handleNamedSequenceGenerator() {
 		final InFlightMetadataCollector metadataCollector = buildingContext.getMetadataCollector();
 
 		final SequenceGeneratorRegistration globalMatch =
@@ -152,16 +105,8 @@ public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver 
 		);
 	}
 
-	private void handleTableStrategy(GeneratedValue generatedValue) {
-		if ( generatedValue.generator().isEmpty() ) {
-			handleUnnamedTableGenerator();
-		}
-		else {
-			handleNamedTableGenerator( generatedValue );
-		}
-	}
-
-	private void handleUnnamedTableGenerator() {
+	@Override
+	protected void handleUnnamedTableGenerator() {
 		final InFlightMetadataCollector metadataCollector = buildingContext.getMetadataCollector();
 
 		final TableGeneratorRegistration globalMatch =
@@ -189,7 +134,8 @@ public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver 
 		);
 	}
 
-	private void handleNamedTableGenerator(GeneratedValue generatedValue) {
+	@Override
+	protected void handleNamedTableGenerator() {
 		final InFlightMetadataCollector metadataCollector = buildingContext.getMetadataCollector();
 
 		final TableGeneratorRegistration globalMatch =
@@ -218,10 +164,17 @@ public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver 
 		);
 	}
 
-	private void handleAutoStrategy(GeneratedValue generatedValue) {
-		final String generator = generatedValue.generator();
-		final String globalRegistrationName = generator.isEmpty() ? entityMapping.getJpaEntityName() : generator;
+	@Override
+	protected void handleUnnamedAutoGenerator() {
+		handleAutoGenerator( entityMapping.getJpaEntityName() );
+	}
 
+	@Override
+	protected void handleNamedAutoGenerator() {
+		handleAutoGenerator( generatedValue.generator() );
+	}
+
+	private void handleAutoGenerator(String globalRegistrationName) {
 		final InFlightMetadataCollector metadataCollector = buildingContext.getMetadataCollector();
 		final GlobalRegistrations globalRegistrations = metadataCollector.getGlobalRegistrations();
 
@@ -265,6 +218,10 @@ public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver 
 			return;
 		}
 
+		if ( handleAsMetaAnnotated() ) {
+			return;
+		}
+
 		// Implicit handling of UUID generation
 		if ( idMember.getType().isImplementor( UUID.class )
 				|| idMember.getType().isImplementor( String.class ) ) {
@@ -272,94 +229,16 @@ public class StrictIdGeneratorResolverSecondPass implements IdGeneratorResolver 
 			return;
 		}
 
-
-		// Handle a few legacy Hibernate generators...
-		if ( !generator.isEmpty() ) {
-			final Class<? extends Generator> legacyNamedGenerator = mapLegacyNamedGenerator( generator, idValue );
-			if ( legacyNamedGenerator != null ) {
-				final Map<String,String> configuration = buildLegacyGeneratorConfig();
-				//noinspection unchecked,rawtypes
-				GeneratorBinder.createGeneratorFrom(
-						new IdentifierGeneratorDefinition( generator, legacyNamedGenerator.getName(), configuration ),
-						idValue,
-						(Map) configuration,
-						buildingContext
-				);
-				return;
-			}
+		if ( handleAsLegacyGenerator() ) {
+			return;
 		}
 
 		handleSequenceGenerator(
 				globalRegistrationName,
-				new SequenceGeneratorJpaAnnotation( generator, metadataCollector.getSourceModelBuildingContext() ),
+				new SequenceGeneratorJpaAnnotation( generatedValue.generator(), metadataCollector.getSourceModelBuildingContext() ),
 				idValue,
 				idMember,
 				buildingContext
 		);
-	}
-
-	private HashMap<String, String> buildLegacyGeneratorConfig() {
-		final Database database = buildingContext.getMetadataCollector().getDatabase();
-		final Dialect dialect = database.getDialect();
-
-		final HashMap<String, String> configuration = new HashMap<>();
-
-		final String tableName = idValue.getTable().getQuotedName( dialect );
-		configuration.put( TABLE, tableName );
-
-		final Column idColumn = (Column) idValue.getSelectables().get( 0);
-		final String idColumnName = idColumn.getQuotedName( dialect );
-		configuration.put( PK, idColumnName );
-
-		configuration.put( ENTITY_NAME, entityMapping.getEntityName() );
-		configuration.put( JPA_ENTITY_NAME, entityMapping.getJpaEntityName() );
-
-		// The table name is not really a good default for subselect entities,
-		// so use the JPA entity name which is short
-		configuration.put(
-				IMPLICIT_NAME_BASE,
-				idValue.getTable().isSubselect()
-						? entityMapping.getJpaEntityName()
-						: idValue.getTable().getName()
-		);
-
-		configuration.put( TABLES, identityTablesString( dialect, entityMapping.getRootClass() ) );
-
-		return configuration;
-	}
-
-	public static void handleSequenceGenerator(
-			String nameFromGeneratedValue,
-			SequenceGenerator generatorAnnotation,
-			SimpleValue idValue,
-			MemberDetails idMember,
-			MetadataBuildingContext buildingContext) {
-		idValue.setCustomIdGeneratorCreator( (creationContext) -> {
-			final BeanContainer beanContainer = GeneratorBinder.beanContainer( buildingContext );
-			final SequenceStyleGenerator identifierGenerator = GeneratorBinder.instantiateGenerator(
-					beanContainer,
-					SequenceStyleGenerator.class
-			);
-			GeneratorAnnotationHelper.prepareForUse(
-					identifierGenerator,
-					generatorAnnotation,
-					idMember,
-					properties -> {
-						if ( generatorAnnotation != null ) {
-							properties.put( GENERATOR_NAME, generatorAnnotation.name() );
-						}
-						else if ( nameFromGeneratedValue != null ) {
-							properties.put( GENERATOR_NAME, nameFromGeneratedValue );
-						}
-						// we need to better handle default allocation-size here...
-						properties.put( INCREMENT_PARAM, fallbackAllocationSize( generatorAnnotation, buildingContext ) );
-					},
-					generatorAnnotation == null
-							? null
-							: (a, properties) -> SequenceStyleGenerator.applyConfiguration( generatorAnnotation, properties::put ),
-					creationContext
-			);
-			return identifierGenerator;
-		} );
 	}
 }
